+++ Tests/Get-PASSession.Tests.ps1
@@ -35,9 +35,7 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 
 	InModuleScope $(Split-Path (Split-Path (Split-Path -Parent $PSCommandPath) -Parent) -Leaf ) {
 		BeforeEach {
-			Mock Get-PASLoggedOnUser -MockWith {
-				[PSCustomObject]@{'Username' = 'SomeUser'; 'Prop2' = 'Val2' }
-			}
+
 
 			$response = Get-PASSession
 		}
@@ -51,33 +49,13 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 
 			It 'has output with expected number of properties' {
 
-				($response | Get-Member -MemberType NoteProperty).length | Should -Be 4
+				($response | Get-Member -MemberType NoteProperty).length | Should -Be 3
 
 			}
 
 			It 'outputs object with expected typename' {
 
 				$response | Get-Member | Select-Object -ExpandProperty typename -Unique | Should -Be psPAS.CyberArk.Vault.Session
-
-			}
-
-			It 'does not throw if Get-PASLoggedOnUser fails' {
-				Mock Get-PASLoggedOnUser -MockWith {
-					throw 'Some Error'
-				}
-
-				{ Get-PASSession } | Should -Not -Throw
-
-			}
-
-			It 'does provides output if Get-PASLoggedOnUser fails' {
-				Mock Get-PASLoggedOnUser -MockWith {
-					throw 'Some Error'
-				}
-
-				$response = Get-PASSession
-
-				($response | Get-Member -MemberType NoteProperty).length | Should -Be 4
 
 			}