+++ psPAS/Functions/Monitoring/Get-PASPSMSession.ps1
@@ -35,13 +35,6 @@ function Get-PASPSMSession {
 			ValueFromPipelinebyPropertyName = $true,
 			ParameterSetName = 'byQuery'
 		)]
-		[int]$Offset,
-
-		[parameter(
-			Mandatory = $false,
-			ValueFromPipelinebyPropertyName = $true,
-			ParameterSetName = 'byQuery'
-		)]
 		[string]$Search,
 
 		[parameter(
@@ -56,14 +49,14 @@ function Get-PASPSMSession {
 			ValueFromPipelinebyPropertyName = $true,
 			ParameterSetName = 'byQuery'
 		)]
-		[int]$FromTime,
+		[datetime]$FromTime,
 
 		[parameter(
 			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true,
 			ParameterSetName = 'byQuery'
 		)]
-		[int]$ToTime,
+		[datetime]$ToTime,
 
 		[parameter(
 			Mandatory = $false,
@@ -97,6 +90,25 @@ function Get-PASPSMSession {
 				#Get Parameters to include in request
 				$boundParameters = $PSBoundParameters | Get-PASParameter
 
+				switch ($PSBoundParameters) {
+
+					{ $PSItem.ContainsKey('FromTime') } {
+
+						$boundParameters['FromTime'] = $FromTime | ConvertTo-UnixTime
+
+					}
+
+					{ $PSItem.ContainsKey('ToTime') } {
+
+						$boundParameters['ToTime'] = $ToTime | ConvertTo-UnixTime
+
+					}
+
+					{ $PSBoundParameters.Keys -notcontains 'Limit' } {
+						$Limit = 25   #default if you call the API with no value
+					}
+
+				}
 				#Create Query String, escaped for inclusion in request URL
 				$queryString = $boundParameters | ConvertTo-QueryString
 
@@ -116,10 +128,47 @@ function Get-PASPSMSession {
 		#send request to PAS web service
 		$result = Invoke-PASRestMethod -Uri $URI -Method GET -WebSession $Script:WebSession
 
-		If ($null -ne $result) {
+		$Total = ($result.LiveSessions).Count
+
+		If ($Total -gt 0) {
+
+			#Set events as output collection
+			$LiveSessions = [Collections.Generic.List[Object]]::New(@($result.LiveSessions))
+
+			#Split Request URL into baseURI & any query string value
+			$URLString = $URI.Split('?')
+			$URI = $URLString[0]
+			$queryString = $URLString[1]
+
+			For ( $Offset = $Limit ; $Limit -eq $Total ; $Offset += $Limit ) {
+
+				#While more risk events to return, create nextLink query value
+				$nextLink = "OffSet=$Offset"
+
+				if ($null -ne $queryString) {
+
+					#If original request contained a queryString, concatenate with nextLink value.
+					$nextLink = "$queryString&$nextLink"
+
+				}
+
+				$result = (Invoke-PASRestMethod -Uri "$URI`?$nextLink" -Method GET -WebSession $Script:WebSession).LiveSessions
+
+				$Total = $result.Count
+
+				#Request nextLink. Add Risk Events to output collection.
+				$Null = $LiveSessions.AddRange($result)
+
+			}
+
+			$Output = $LiveSessions
+
+		}
+
+		If ($null -ne $Output) {
 
 			#Return Results
-			$result.LiveSessions | Add-ObjectDetail -typename psPAS.CyberArk.Vault.PSM.Session
+			$Output | Add-ObjectDetail -typename psPAS.CyberArk.Vault.PSM.Session
 
 		} #process