+++ CHANGELOG.md
@@ -11,10 +11,26 @@
 - N/A
 
 ### Updated
-- N/A
+- `Get-PASPSMRecording`
+  - Removes `Offset` Parameter
+  - Updates `FromTime` & `ToTime` parameters to `[datetime]` types
+  - Returns all pages of results instead of only the first page of results
+- `Get-PASPSMSession`
+  - Removes `Offset` Parameter
+  - Updates `FromTime` & `ToTime` parameters to `[datetime]` types
+  - Returns all pages of results instead of only the first page of results
+- `Get-PASAccount`
+  - Removes `Offset` Parameter
+- `Get-PASDiscoveredAccount`
+  - Removes `Offset` Parameter
 
 ### Fixed
-- N/A
+- `Get-PASSession`
+  - Removes `UserName` from command output, avoiding error condition on expired session.
+- `Get-PASPlatform`
+  - Adds `search` parameter to the default `targets` parameterset
+- ISPSS Error Handling
+  - Fixes issue where error returned from ISPSS solution may not be handled properly
 
 ## **6.1.50**